Remove duplicate logic in DummyActionComponentExample
