+++ Demo/Common/IntegrationExamples/Actions/ActionComponentExample.swift
@@ -8,28 +8,37 @@ import Adyen
 import AdyenActions
 import AdyenCheckout
 
-internal final class DummyActionComponentExample: InitialDataAdvancedFlowProtocol {
-    
+/// Standalone example that presents an action's UI without going through a full payment flow.
+///
+/// It decodes a provided action JSON string and hands it to an `ActionOnlyCheckout`, which internally
+/// routes it to the matching component (e.g. `VoucherComponent`, `QRCodeComponent`). This is useful for
+/// quickly iterating on action UIs. The dummy action to present is provided by the caller.
+internal final class ActionComponentExample: InitialDataAdvancedFlowProtocol {
+
     internal weak var presenter: PresenterExampleProtocol?
-    
+
     private var checkout: ActionOnlyCheckout?
-    
+
     internal lazy var apiClient = ApiClientHelper.generateApiClient()
     private lazy var asyncApiClient = ApiClientHelper.generateAsyncApiClient()
     /// comes from demo app protocol, unused on new structure
     internal var context: AdyenContext?
 
-    internal init() {}
-    
+    private let actionString: String
+
+    internal init(actionString: String) {
+        self.actionString = actionString
+    }
+
     internal func start() {
         startLoading()
-        
+
         Task { @MainActor in
             do {
                 let checkout = try await createCheckout()
                 let actionData = actionString.data(using: .utf8)
                 let action = try JSONDecoder().decode(Action.self, from: actionData!)
-                
+
                 hideLoading()
                 self.checkout = checkout
                 checkout.handle(action: action)
@@ -39,7 +48,7 @@ internal final class DummyActionComponentExample: InitialDataAdvancedFlowProtoco
             }
         }
     }
-    
+
     private func createCheckout() async throws -> ActionOnlyCheckout {
         let configuration = try CheckoutConfiguration(
             environment: ConfigurationConstants.componentsEnvironment,
@@ -49,7 +58,7 @@ internal final class DummyActionComponentExample: InitialDataAdvancedFlowProtoco
                 isEnabled: ConfigurationConstants.current.analyticsSettings.isEnabled
             )
         ) {}
-        
+
         return try await Checkout.setup(
             configuration: configuration,
             presentationDelegate: self
@@ -76,11 +85,11 @@ internal final class DummyActionComponentExample: InitialDataAdvancedFlowProtoco
             return .completion(resultCode: "Error")
         }
     }
-    
+
     private func startLoading() {
         presenter?.showLoadingIndicator()
     }
-    
+
     @MainActor
     private func handleError(_ error: Error) {
         presenter?.presentAlert(withTitle: "Error", message: error.localizedDescription)
@@ -90,28 +99,27 @@ internal final class DummyActionComponentExample: InitialDataAdvancedFlowProtoco
     private func hideLoading() {
         presenter?.hideLoadingIndicator()
     }
-    
+
     private func dismissAndShowAlert(_ success: Bool, _ message: String) {
         presenter?.dismiss {
-            // Payment is processed. Add your code here.
             let title = success ? "Success" : "Error"
             self.presenter?.presentAlert(withTitle: title, message: message)
         }
     }
-    
-    private let actionString = """
-          {
-            "paymentMethodType" : "pix",
-            "paymentData" : "paymentData",
-            "qrCodeData" : "TestQRCodeEMVToken",
-            "type" : "qrCode"
-          }
-    """
 }
 
-extension DummyActionComponentExample: PresentationDelegate {
-    
+extension ActionComponentExample: PresentationDelegate {
+
     func present(viewController: UIViewController) {
-        presenter?.present(viewController: viewController, completion: nil)
+        // Wrap in a navigation controller with a cancel button so the action can be dismissed
+        // when testing it in isolation (the standalone Checkout flow adds no navigation chrome).
+        let navigationController = UINavigationController(rootViewController: viewController)
+        viewController.navigationItem.leftBarButtonItem = UIBarButtonItem(
+            systemItem: .cancel,
+            primaryAction: UIAction { [weak self] _ in
+                self?.presenter?.dismiss(completion: nil)
+            }
+        )
+        presenter?.present(viewController: navigationController, completion: nil)
     }
 }
